Time: 7 ms (40.07%) | Memory: 71.6 MB (70.75%) - LeetSync

+++ 878-shifting-letters/shifting-letters.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     string shiftingLetters(string s, vector<int>& shifts) {
         int n = s.size();
-           long long tot = 0;
+           int tot = 0;
 
         for(int i = n-1; i>=0; i--){
             tot = (tot + shifts[i]) % 26;
